+++ examples/atm_mib/src/agentpp_notifytest_mib.cpp
@@ -88,7 +88,7 @@ int agentppNotifyTest::set(const Vbx& vb)
 {
 
 	//--AgentGen BEGIN=agentppNotifyTest::set
-	long v = 0;
+	int32_t v = 0;
 	vb.get_value(v);
 	switch (v) {
 	case e_agentppNotifyTestAllTypes:
@@ -101,7 +101,7 @@ int agentppNotifyTest::set(const Vbx& vb)
 
 bool agentppNotifyTest::value_ok(const Vbx& vb)
 {
-	long v = 0;
+	int32_t v = 0;
 	vb.get_value(v);
 	if ((v != 1)
 	    ) return FALSE;
