+++ examples/atm_mib/src/agentpp_test_mib.cpp
@@ -102,7 +102,7 @@ void agentppTestTimeout::set_state(unsigned long l)
 int agentppTestTimeout::set(const Vbx& vb)
 {
 	//--AgentGen BEGIN=agentppTestTimeout::set
-	unsigned long timeout = 0;
+	uint32_t timeout = 0;
 	vb.get_value(timeout);
 	struct timeval tvptr;
 	tvptr.tv_sec  = timeout/1000;    // wait up to sec seconds
@@ -115,7 +115,7 @@ int agentppTestTimeout::set(const Vbx& vb)
 
 bool agentppTestTimeout::value_ok(const Vbx& vb)
 {
-	unsigned long v = 0;
+	uint32_t v = 0;
 	vb.get_value(v);
 	if (!((v <= 1000000ul)))
 		 return FALSE;
@@ -380,14 +380,14 @@ void agentppTestSparseCol1::get_request(Request* req, int ind)
 	MibLeaf::get_request(req, ind);
 }
 
-long agentppTestSparseCol1::get_state()
+int32_t agentppTestSparseCol1::get_state()
 {
 	//--AgentGen BEGIN=agentppTestSparseCol1::get_state
 	//--AgentGen END
-	return (long)*((SnmpInt32*)value);
+	return (int32_t)*((SnmpInt32*)value);
 }
 	
-void agentppTestSparseCol1::set_state(long l)
+void agentppTestSparseCol1::set_state(int32_t l)
 {
 	//--AgentGen BEGIN=agentppTestSparseCol1::set_state
 	//--AgentGen END
@@ -645,14 +645,14 @@ MibEntryPtr agentppTestSparseRowStatus::clone()
 	return other;
 }
 
-long agentppTestSparseRowStatus::get_state()
+int32_t agentppTestSparseRowStatus::get_state()
 {
 	//--AgentGen BEGIN=agentppTestSparseRowStatus::get_state
 	//--AgentGen END
-	return (long)*((SnmpInt32*)value);
+	return (int32_t)*((SnmpInt32*)value);
 }
 	
-void agentppTestSparseRowStatus::set_state(long l)
+void agentppTestSparseRowStatus::set_state(int32_t l)
 {
 	//--AgentGen BEGIN=agentppTestSparseRowStatus::set_state
 	//--AgentGen END
