+++ src/v3_mib.cpp
@@ -152,7 +152,7 @@ UsmUserTableStatus::~UsmUserTableStatus()
 int UsmUserTableStatus::set(const Vbx& vb)
 {
   undo = value->clone();
-  long rs = 0;
+  int32_t rs = 0;
   if (vb.get_value(rs) != SNMP_CLASS_SUCCESS)
       return SNMP_ERROR_WRONG_TYPE;
